backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
cassandra-driver==3.29.0
pydantic==2.5.0
python-multipart==0.0.6
websockets==12.0
python-dateutil==2.8.2
pandas==2.1.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
kafka-python==2.0.2 